requirements.txt
streamlit
numpy
opencv-python
tensorflow
Pillow
matplotlib
lime
scikit-image

